Major rewrite of the publish part of the plugin.
The directory has been changed to /var/www/pub, to match the other plugins,
and to make selinux happy.

The packages are symlinked under a 'repo' subdirectory, since I envinsion the
need for a metadata (in json format) file, that lists metadata about the
MSI/EXE files.

Using steps makes things a lot easier to process.